🔨 Refactor import paths for consistency in tests
